Add factories for stages and scope handlers
