+++ packages/cursorless-engine/src/processTargets/MarkStageFactory.ts
@@ -1,0 +1,6 @@
+import { Mark } from "@cursorless/common";
+import { MarkStage } from "./PipelineStages.types";
+
+export interface MarkStageFactory {
+  create(mark: Mark): MarkStage;
+}
