+++ packages/cursorless-engine/src/processTargets/index.ts
@@ -1,2 +1,1 @@
-import processTargets from "./processTargets";
-export default processTargets;
+export * from "./TargetPipeline";
