+++ packages/cursorless-engine/src/processTargets/marks/RangeMarkStage.ts
@@ -1,16 +1,19 @@
-import { Target } from "../../typings/target.types";
 import { RangeMark } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
-import getMarkStage from "../getMarkStage";
+import { Target } from "../../typings/target.types";
+import { MarkStageFactory } from "../MarkStageFactory";
 import { MarkStage } from "../PipelineStages.types";
-import { targetsToContinuousTarget } from "../processTargets";
+import { targetsToContinuousTarget } from "../TargetPipeline";
 
 export default class RangeMarkStage implements MarkStage {
-  constructor(private mark: RangeMark) {}
+  constructor(
+    private markStageFactory: MarkStageFactory,
+    private mark: RangeMark,
+  ) {}
 
   run(context: ProcessedTargetsContext): Target[] {
-    const anchorStage = getMarkStage(this.mark.anchor);
-    const activeStage = getMarkStage(this.mark.active);
+    const anchorStage = this.markStageFactory.create(this.mark.anchor);
+    const activeStage = this.markStageFactory.create(this.mark.active);
     const anchorTargets = anchorStage.run(context);
     const activeTargets = activeStage.run(context);
