+++ packages/cursorless-engine/src/processTargets/modifiers/InteriorStage.ts
@@ -1,27 +1,36 @@
-import { Target } from "../../typings/target.types";
 import {
   ExcludeInteriorModifier,
   InteriorOnlyModifier,
 } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
+import { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import { ModifierStage } from "../PipelineStages.types";
-import { containingSurroundingPairIfUntypedStage } from "./commonContainingScopeIfUntypedStages";
+import { containingSurroundingPairIfUntypedModifier } from "./commonContainingScopeIfUntypedModifiers";
 
 export class InteriorOnlyStage implements ModifierStage {
-  constructor(private modifier: InteriorOnlyModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: InteriorOnlyModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    return containingSurroundingPairIfUntypedStage
+    return this.modifierStageFactory
+      .create(containingSurroundingPairIfUntypedModifier)
       .run(context, target)
       .flatMap((target) => target.getInteriorStrict());
   }
 }
 
 export class ExcludeInteriorStage implements ModifierStage {
-  constructor(private modifier: ExcludeInteriorModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: ExcludeInteriorModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    return containingSurroundingPairIfUntypedStage
+    return this.modifierStageFactory
+      .create(containingSurroundingPairIfUntypedModifier)
       .run(context, target)
       .flatMap((target) => target.getBoundaryStrict());
   }
