+++ packages/cursorless-engine/src/processTargets/modifiers/LeadingTrailingStages.ts
@@ -1,8 +1,9 @@
-import { Target } from "../../typings/target.types";
 import { LeadingModifier, TrailingModifier } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
+import { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import { ModifierStage } from "../PipelineStages.types";
-import { containingTokenIfUntypedStage } from "./commonContainingScopeIfUntypedStages";
+import { containingTokenIfUntypedModifier } from "./commonContainingScopeIfUntypedModifiers";
 
 /**
  * Throw this error if user has requested leading or trailing delimiter but no
@@ -16,29 +17,41 @@ class NoDelimiterError extends Error {
 }
 
 export class LeadingStage implements ModifierStage {
-  constructor(private modifier: LeadingModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: LeadingModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    return containingTokenIfUntypedStage.run(context, target).map((target) => {
-      const leading = target.getLeadingDelimiterTarget();
-      if (leading == null) {
-        throw new NoDelimiterError("leading");
-      }
-      return leading;
-    });
+    return this.modifierStageFactory
+      .create(containingTokenIfUntypedModifier)
+      .run(context, target)
+      .map((target) => {
+        const leading = target.getLeadingDelimiterTarget();
+        if (leading == null) {
+          throw new NoDelimiterError("leading");
+        }
+        return leading;
+      });
   }
 }
 
 export class TrailingStage implements ModifierStage {
-  constructor(private modifier: TrailingModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: TrailingModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    return containingTokenIfUntypedStage.run(context, target).map((target) => {
-      const trailing = target.getTrailingDelimiterTarget();
-      if (trailing == null) {
-        throw new NoDelimiterError("trailing");
-      }
-      return trailing;
-    });
+    return this.modifierStageFactory
+      .create(containingTokenIfUntypedModifier)
+      .run(context, target)
+      .map((target) => {
+        const trailing = target.getTrailingDelimiterTarget();
+        if (trailing == null) {
+          throw new NoDelimiterError("trailing");
+        }
+        return trailing;
+      });
   }
 }
