+++ packages/cursorless-engine/src/processTargets/modifiers/OrdinalScopeStage.ts
@@ -1,6 +1,7 @@
-import { Target } from "../../typings/target.types";
 import { OrdinalScopeModifier } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
+import { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import { ModifierStage } from "../PipelineStages.types";
 import {
   createRangeTargetFromIndices,
@@ -8,10 +9,14 @@ import {
 } from "./targetSequenceUtils";
 
 export class OrdinalScopeStage implements ModifierStage {
-  constructor(private modifier: OrdinalScopeModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: OrdinalScopeModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
     const targets = getEveryScopeTargets(
+      this.modifierStageFactory,
       context,
       target,
       this.modifier.scopeType,
