+++ packages/cursorless-engine/src/processTargets/modifiers/RangeModifierStage.ts
@@ -1,16 +1,19 @@
-import { Target } from "../../typings/target.types";
 import { RangeModifier } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
-import getModifierStage from "../getModifierStage";
+import { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import { ModifierStage } from "../PipelineStages.types";
-import { targetsToContinuousTarget } from "../processTargets";
+import { targetsToContinuousTarget } from "../TargetPipeline";
 
 export default class RangeModifierStage implements ModifierStage {
-  constructor(private modifier: RangeModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private modifier: RangeModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    const anchorStage = getModifierStage(this.modifier.anchor);
-    const activeStage = getModifierStage(this.modifier.active);
+    const anchorStage = this.modifierStageFactory.create(this.modifier.anchor);
+    const activeStage = this.modifierStageFactory.create(this.modifier.active);
     const anchorTargets = anchorStage.run(context, target);
     const activeTargets = activeStage.run(context, target);
