+++ packages/cursorless-engine/src/processTargets/modifiers/RelativeInclusiveScopeStage.ts
@@ -1,17 +1,18 @@
-import { NoContainingScopeError, Range, TextEditor } from "@cursorless/common";
 import type { Direction, RelativeScopeModifier } from "@cursorless/common";
-import type { Target } from "../../typings/target.types";
+import { NoContainingScopeError, Range, TextEditor } from "@cursorless/common";
 import type { ProcessedTargetsContext } from "../../typings/Types";
+import type { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import type { ModifierStage } from "../PipelineStages.types";
+import { TooFewScopesError } from "./TooFewScopesError";
 import { constructScopeRangeTarget } from "./constructScopeRangeTarget";
 import { getContainingScope } from "./getContainingScope";
 import { runLegacy } from "./relativeScopeLegacy";
-import getScopeHandler from "./scopeHandlers/getScopeHandler";
+import { ScopeHandlerFactory } from "./scopeHandlers/ScopeHandlerFactory";
 import getScopeRelativeToPosition from "./scopeHandlers/getScopeRelativeToPosition";
 import getScopesOverlappingRange from "./scopeHandlers/getScopesOverlappingRange";
 import type { TargetScope } from "./scopeHandlers/scope.types";
 import type { ScopeHandler } from "./scopeHandlers/scopeHandler.types";
-import { TooFewScopesError } from "./TooFewScopesError";
 
 /**
  * Handles relative modifiers that include targets intersecting with the input,
@@ -37,16 +38,25 @@ import { TooFewScopesError } from "./TooFewScopesError";
  *    direction is backward.
  */
 export class RelativeInclusiveScopeStage implements ModifierStage {
-  constructor(private modifier: RelativeScopeModifier) {}
+  constructor(
+    private modifierStageFactory: ModifierStageFactory,
+    private scopeHandlerFactory: ScopeHandlerFactory,
+    private modifier: RelativeScopeModifier,
+  ) {}
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
-    const scopeHandler = getScopeHandler(
+    const scopeHandler = this.scopeHandlerFactory.create(
       this.modifier.scopeType,
       target.editor.document.languageId,
     );
 
     if (scopeHandler == null) {
-      return runLegacy(this.modifier, context, target);
+      return runLegacy(
+        this.modifierStageFactory,
+        this.modifier,
+        context,
+        target,
+      );
     }
 
     const { isReversed, editor, contentRange: inputRange } = target;
