+++ packages/cursorless-engine/src/processTargets/modifiers/RelativeScopeStage.ts
@@ -1,9 +1,11 @@
-import type { Target } from "../../typings/target.types";
 import type { RelativeScopeModifier } from "@cursorless/common";
 import type { ProcessedTargetsContext } from "../../typings/Types";
+import type { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 import type { ModifierStage } from "../PipelineStages.types";
 import RelativeExclusiveScopeStage from "./RelativeExclusiveScopeStage";
 import { RelativeInclusiveScopeStage } from "./RelativeInclusiveScopeStage";
+import { ScopeHandlerFactory } from "./scopeHandlers/ScopeHandlerFactory";
 
 /**
  * Implements relative scope modifiers like "next funk", "two tokens", etc.
@@ -14,11 +16,23 @@ import { RelativeInclusiveScopeStage } from "./RelativeInclusiveScopeStage";
  */
 export default class RelativeScopeStage implements ModifierStage {
   private modiferStage: ModifierStage;
-  constructor(private modifier: RelativeScopeModifier) {
+  constructor(
+    modifierStageFactory: ModifierStageFactory,
+    scopeHandlerFactory: ScopeHandlerFactory,
+    modifier: RelativeScopeModifier,
+  ) {
     this.modiferStage =
-      this.modifier.offset === 0
-        ? new RelativeInclusiveScopeStage(modifier)
-        : new RelativeExclusiveScopeStage(modifier);
+      modifier.offset === 0
+        ? new RelativeInclusiveScopeStage(
+            modifierStageFactory,
+            scopeHandlerFactory,
+            modifier,
+          )
+        : new RelativeExclusiveScopeStage(
+            modifierStageFactory,
+            scopeHandlerFactory,
+            modifier,
+          );
   }
 
   run(context: ProcessedTargetsContext, target: Target): Target[] {
