+++ packages/cursorless-engine/src/processTargets/modifiers/commonContainingScopeIfUntypedModifiers.ts
@@ -1,4 +1,4 @@
-import { ModifyIfUntypedStage } from "./ConditionalModifierStages";
+import { Modifier } from "@cursorless/common";
 // NB: We import `Target` below just so that @link below resolves.  Once one of
 // the following issues are fixed, we can either remove the above line or
 // switch to `{import("foo")}` syntax in the `{@link}` tag.
@@ -12,36 +12,34 @@ import type { Target } from "../../typings/target.types";
  * target has no explicit scope type, ie if {@link Target.hasExplicitScopeType}
  * is `false`.
  */
-export const containingSurroundingPairIfUntypedStage = new ModifyIfUntypedStage(
-  {
-    type: "modifyIfUntyped",
-    modifier: {
-      type: "containingScope",
-      scopeType: { type: "surroundingPair", delimiter: "any" },
-    },
+export const containingSurroundingPairIfUntypedModifier: Modifier = {
+  type: "modifyIfUntyped",
+  modifier: {
+    type: "containingScope",
+    scopeType: { type: "surroundingPair", delimiter: "any" },
   },
-);
+};
 
 /**
  * Expands the given target to the nearest containing line if the target has no
  * explicit scope type, ie if {@link Target.hasExplicitScopeType} is `false`.
  */
-export const containingLineIfUntypedStage = new ModifyIfUntypedStage({
+export const containingLineIfUntypedModifier: Modifier = {
   type: "modifyIfUntyped",
   modifier: {
     type: "containingScope",
     scopeType: { type: "line" },
   },
-});
+};
 
 /**
  * Expands the given target to the nearest containing token if the target has no
  * explicit scope type, ie if {@link Target.hasExplicitScopeType} is `false`.
  */
-export const containingTokenIfUntypedStage = new ModifyIfUntypedStage({
+export const containingTokenIfUntypedModifier: Modifier = {
   type: "modifyIfUntyped",
   modifier: {
     type: "containingScope",
     scopeType: { type: "token" },
   },
-});
+};
