+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/ScopeHandlerFactory.ts
@@ -1,0 +1,6 @@
+import type { ScopeType } from "@cursorless/common";
+import type { ScopeHandler } from "./scopeHandler.types";
+
+export interface ScopeHandlerFactory {
+  create(scopeType: ScopeType, languageId: string): ScopeHandler | undefined;
+}
