+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/index.ts
@@ -16,5 +16,4 @@ export * from "./OneOfScopeHandler";
 export { default as OneOfScopeHandler } from "./OneOfScopeHandler";
 export * from "./ParagraphScopeHandler";
 export { default as ParagraphScopeHandler } from "./ParagraphScopeHandler";
-export * from "./getScopeHandler";
-export { default as getScopeHandler } from "./getScopeHandler";
+export * from "./ScopeHandlerFactoryImpl";
