+++ packages/cursorless-engine/src/processTargets/modifiers/targetSequenceUtils.ts
@@ -1,7 +1,7 @@
-import { Target } from "../../typings/target.types";
 import { ScopeType } from "@cursorless/common";
 import { ProcessedTargetsContext } from "../../typings/Types";
-import getModifierStage from "../getModifierStage";
+import { Target } from "../../typings/target.types";
+import { ModifierStageFactory } from "../ModifierStageFactory";
 
 export class OutOfRangeError extends Error {
   constructor() {
@@ -42,11 +42,12 @@ export function createRangeTargetFromIndices(
 }
 
 export function getEveryScopeTargets(
+  modifierStageFactory: ModifierStageFactory,
   context: ProcessedTargetsContext,
   target: Target,
   scopeType: ScopeType,
 ): Target[] {
-  const containingStage = getModifierStage({
+  const containingStage = modifierStageFactory.create({
     type: "everyScope",
     scopeType,
   });
